Refactor to hibernate: admin(delete, wa mutual delete), user (profile update)

+++ web-module/src/test/java/by/tut/mdcatalog/project2/web/controller/ReviewControllerIntegrationTest.java
@@ -47,7 +47,7 @@ public class ReviewControllerIntegrationTest {
     @Test
     @WithMockUser(authorities = {AuthorizationConstants.ADMIN_ROLE_NAME})
     public void shouldDeleteReviewsForAdmin() {
-        int[] ids = {2, 3};
+        Long[] ids = {2L, 3L};
         String url = reviewController.deleteReviews(ids);
         Assert.assertEquals("redirect:/success", url);
     }
